modeling/requirements.txt
numpy>=2.1,<2.5
pandas>=2.2
PyYAML>=6.0
scikit-learn>=1.4
rank-bm25>=0.2.2
tqdm>=4.66
transformers>=4.40
torch>=2.2
sentence-transformers>=2.7
setuptools<82
